pyproject.toml
[tool.poetry]
name = "task-27"
version = "0.1.0"
description = ""
authors = ["HAMarat <smile56680@gmail.com>"]
readme = "README.md"
packages = [{include = "task_27"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
pillow = "^9.5.0"
psycopg2 = "^2.9.6"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
python-dateutil = "^2.8.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
